pyproject.toml
[project]
name = "recsysfood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "mypy>=1.18.2",
    "networkx>=3.6",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "scikit-learn>=1.7.2",
    "supabase>=2.24.0",
    "torch>=2.9.1",
    "torch-geometric>=2.7.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
    "pandas-stubs>=2.3.3",
    "types-networkx>=3.6.0",
    "types-pytz>=2025.1.0",
    "types-python-dateutil>=2.9.0",
    "types-pytz>=2025.1.0",
    "schedule>=1.2.2",
    "imblearn>=0.0",
    "transformers>=4.57.3",
    "sentence-transformers>=5.1.2",
    "shap>=0.49.1",
    "lime>=0.2.0.1",
    "nltk>=3.9.2",
]
